converted the crc into c++ with namespace.
Since it conflicts with other crc implementation in imagemagic lib.
This resulted in png images that would not open due to "crc error".
