+++ Makefile.am
@@ -10,8 +10,7 @@ endif
 odr_padenc_CXXFLAGS = $(GITVERSION_FLAGS) @MAGICKWAND_CFLAGS@ -Icontrib -Wall -Wextra
 odr_padenc_LDADD    = @MAGICKWAND_LDADD@
 odr_padenc_SOURCES  = src/odr-padenc.cpp \
-					  src/crc.h \
-					  src/crc.c
+					  src/crc.cpp
 
 bin_PROGRAMS = odr-padenc$(EXEEXT)
